Deprecate low-level frame and DFA imports
